+++ app/src/main/java/com/lateinit/rightweight/data/api/UserApiService.kt
@@ -8,7 +8,6 @@ import com.lateinit.rightweight.data.model.remote.WriteRequestBody
 import com.lateinit.rightweight.data.remote.model.HistoryField
 import com.lateinit.rightweight.data.remote.model.RemoteData
 import com.lateinit.rightweight.data.remote.model.RootField
-import com.lateinit.rightweight.data.remote.model.RoutineField
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.PATCH
@@ -34,10 +33,6 @@ interface UserApiService {
         path: String
     ): DocumentsResponse<DetailResponse<RemoteData>>
 
-    @POST("./documents:runQuery")
-    suspend fun getUserRoutine(
-        @Body query: RunQueryBody
-    ): List<DocumentResponse<RoutineField>>
 
     @POST("documents/user/{userId}/:runQuery")
     suspend fun getLastHistory(
