+++ app/src/main/java/com/lateinit/rightweight/data/datasource/local/RoutineLocalDataSource.kt
@@ -29,6 +29,8 @@ interface RoutineLocalDataSource {
 
     suspend fun getRoutineWithDaysByRoutineId(routineId: String): RoutineWithDays
 
+    suspend fun getAllRoutineWithDays(): List<RoutineWithDays>
+
     fun getAllRoutines(): Flow<List<Routine>>
 
     fun getDayWithExercisesByDayId(dayId: String): Flow<DayWithExercises>
