+++ app/src/main/java/com/lateinit/rightweight/data/datasource/local/impl/RoutineLocalDataSourceImpl.kt
@@ -56,6 +56,10 @@ class RoutineLocalDataSourceImpl @Inject constructor(
         return routineDao.getRoutineWithDaysByRoutineId(routineId)
     }
 
+    override suspend fun getAllRoutineWithDays(): List<RoutineWithDays> {
+        return routineDao.getAllRoutineWithDays()
+    }
+
     override fun getDayWithExercisesByDayId(dayId: String): Flow<DayWithExercises> {
         return routineDao.getDayWithExercisesByDayId(dayId)
     }
